requirements.txt
# FastAPI and ASGI
fastapi==0.104.0
uvicorn==0.24.0
python-multipart==0.0.6

# Pydantic
pydantic==2.4.2
pydantic-settings==2.0.3
email-validator==2.0.0.post2

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.28.0

# Redis and Caching
redis==5.0.1
celery==5.3.4

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Monitoring and Logging
prometheus-client==0.19.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-exporter-otlp-proto-grpc==1.21.0
opentelemetry-semantic-conventions==0.42b0
structlog==23.2.0

# HTTP Client
httpx==0.25.1
tenacity==8.2.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Code Quality
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.0 